backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
selenium==4.15.2
playwright==1.40.0
Appium-Python-Client==3.1.1
requests==2.31.0
Jinja2==3.1.2
Pillow==10.1.0
python-dotenv==1.0.0
psutil==5.9.6
PyYAML==6.0.1
# ASGI/Channels
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
# Django filters
django-filter==23.5
# Message Queue
pika==1.3.2
# Encryption
cryptography==41.0.7
# AI Service (V2.0 LLM Gateway)
httpx>=0.27.0
loguru>=0.7.0
